requirements.txt
py-clob-client>=0.9.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
rapidfuzz>=3.6.0
python-socketio[asyncio_client]>=5.11.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
aioresponses>=0.7.6
